Created TalonObjects subystem; Defined OI() constructor;
Created Button instance variables in OI() class as well as button number
constants for representing the buttons in the Logitech F310 controller;

Created "talons" instance variable in Robot class, as well as mecanum
instance variable.

The difference between these is mecanum is intended for the Game
Controller while talons is intended for commands, since it provides some
macro functions in the TalonObjects class. Write your own as well! Make
Commands!