Created the objects for the two pneumatics in the back of the robot.
Created methods which should in theory allow us to extend the pneumatics and retract them.

Crated class for the pneumatics on the claw mechanism

Created methods for opening and clasing the claw.

Not sure if we will use it but I made a class for the pneumatic we might use to push down the claw during autonomous.

Created methods which we can use to extend and retract the pushPneumatic

Continued working on the actuators class (stuggling a bit: not sure whether to tackle them in the same manner as the wheels)

Worked on the actuators class. Created the objects for the big and small actuators.

Created methods to control the actuators

Finished mapping the subsystems on the robot as of now. All that is left to do is create some commands and set up the correct ports.